Update cl.rb
So I am new at this and looked to find something that would be interesting to me... both to explore github, set up gems and also perhaps to make it do what I want.

After getting this to run, I noticed that nothing was printing in the title.. and that was because the link: has was set to be an HTML link, so when it was rendered in the browser, it showed up as nothing.  

I changed the link to have text include the full url as well as default text in the link.  

Also, I changed the default location, as for some reason, the ones that are in there do not work.